Sir-Zia/basic/agent/extra/sumryTo Text/pyproject.toml
[project]
name = "dataclass"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "openai-agents==0.0.16",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
]
